Move BusinessProduct to Business Package
+ Make Product Fields Protected
