+++ abstraction/src/main/java/us/teaminceptus/novaconomy/util/inventory/Generator.java
@@ -30,7 +30,7 @@ import us.teaminceptus.novaconomy.api.economy.market.Receipt;
 import us.teaminceptus.novaconomy.api.events.business.BusinessAdvertiseEvent;
 import us.teaminceptus.novaconomy.api.player.NovaPlayer;
 import us.teaminceptus.novaconomy.api.settings.Settings;
-import us.teaminceptus.novaconomy.api.util.BusinessProduct;
+import us.teaminceptus.novaconomy.api.business.BusinessProduct;
 
 import java.lang.reflect.Method;
 import java.util.*;
