+++ api/src/main/java/us/teaminceptus/novaconomy/api/events/player/economy/PlayerPurchaseProductEvent.java
@@ -4,7 +4,7 @@ import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import us.teaminceptus.novaconomy.api.business.BusinessStatistics;
-import us.teaminceptus.novaconomy.api.util.BusinessProduct;
+import us.teaminceptus.novaconomy.api.business.BusinessProduct;
 
 /**
  * Called after a Player purchases something
